refractor: Split and refractor utils methods
