+++ src/utils/consolex.ts
@@ -1,0 +1,8 @@
+import { red } from "../../deps.ts";
+
+function error(msg: string) {
+  console.error(`${red("error")} ${msg}`);
+  Deno.exit(1);
+}
+
+export { error };
